Re-added support for theme-specific routes
