+++ src/Wardrobe/Core/Controllers/BaseController.php
@@ -24,8 +24,6 @@ class BaseController extends Controller {
 
 		$this->theme = Config::get('core::wardrobe.theme');
 
-		$this->setupViews();
-
 		$presence = Validator::getPresenceVerifier();
 		$presence->setConnection('wardrobe');
 
@@ -50,14 +48,6 @@ class BaseController extends Controller {
 		if ( ! is_null($this->layout))
 		{
 			$this->layout = View::make($this->layout);
-		}
-	}
-
-	protected function setupViews()
-	{
-		View::addLocation(public_path().'/'.Config::get('core::wardrobe.theme_dir'));
-		foreach (Config::get('core::wardrobe.view_dirs') as $dir) {
-			View::addLocation($dir);
 		}
 	}
